pyproject.toml
[tool.poetry]
name = "flask-quotes"
version = "0.1.0"
description = ""
authors = ["NightSpring1 <deroy193@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
sqlalchemy = "^2.0.20"
flask-sqlalchemy = "^3.0.5"
redis-lru = "^0.1.2"
flask-login = "^0.6.2"
waitress = "^2.1.2"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
